+++ visual-js/visual-playwright/src/fixtures.ts
@@ -1,11 +1,23 @@
-import { TestFixture, TestInfo } from '@playwright/test';
-import { Page } from 'playwright-core';
-import { SauceVisualParams } from './types';
-import { sauceVisualCheck } from './playwright';
+import type { TestFixture, TestInfo } from '@playwright/test';
+import type { Page } from 'playwright-core';
+import type { SauceVisualParams } from './types';
+import { sauceVisualCheck, sauceVisualResults } from './playwright';
+import type { DiffStatus } from '@saucelabs/visual';
 
 export type SauceVisualFixtures = {
   sauceVisual: {
+    /**
+     * Takes a snapshot of the current page and uploads it to Sauce Labs for visual diffing.
+     * @param name
+     * @param options
+     */
     visualCheck: (name: string, options?: SauceVisualParams) => Promise<void>;
+    /**
+     * Returns the visual results for the active test. Can only be used inside a
+     * `test` or `afterEach` block since it uses the current test context for finding matching
+     * visual results.
+     */
+    sauceVisualResults: () => Promise<Record<DiffStatus, number>>;
   };
 };
 
@@ -20,6 +32,7 @@ export const sauceVisualFixtures: (defaultOptions?: SauceVisualParams) => {
           ...options,
         });
       },
+      sauceVisualResults: async () => await sauceVisualResults(testInfo),
     });
   },
 });